# docs/swagger.yaml
basePath: /
definitions:
...
  enum.Role:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - Admin
    - User
...
  models.User:
    description: User represents a user with details such as name, email, password,
      etc.
    properties:
      address:
        description: |-
          Address of the user, optional
          @example "123 Main St, Springfield, IL"
        type: string
      age:
        description: |-
          Age of the user, optional
          @example 25
        type: integer
      created_at:
        description: |-
          Timestamp when the user was created
          @example "2025-04-10T14:22:00Z"
        type: string
      email:
        description: |-
          Email of the user, unique
          @example "johndoe@example.com"
        type: string
      gender:
        description: |-
          Gender of the user, optional
          @example "male"
        type: string
      id:
        description: |-
          ID of the user, automatically generated
          @example 123e4567-e89b-12d3-a456-426614174000
        type: string
      is_verified:
        description: |-
          Whether the user's email has been verified
          @example false
        type: boolean
      name:
        description: |-
          Name of the user
          @example "John Doe"
        type: string
      password:
        description: |-
          Password of the user
          @example "password123"
        type: string
      phone:
        description: |-
          Phone number of the user, optional
          @example "+1234567890"
        type: string
      role:
        allOf:
        - $ref: '#/definitions/enum.Role'
        description: |-
          Role of the user, default is 'user'
          @example "admin"
      updated_at:
        description: |-
          Timestamp when the user was last updated
          @example "2025-04-10T14:22:00Z"
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: QFirst is a robust and easy-to-use booking system designed to help
    businesses manage appointments and reservations efficiently.
  title: QFirst API Documentation
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Add a new user to the platform
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes a user by their unique ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Retrieve a user's details by their unique ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
  CsrfToken:
    description: CSRF protection token.
    in: header
    name: X-CSRF-Token
    type: apiKey
swagger: "2.0"